Changed project from C to Javascript
I removed all files with the old OpenGL code and started again:

The new template features vue, whitestorm and bootstrap Frameworks

Vue shows the website. It will replace several subsites inside the
main content view according the menu, that yet has to be programmed.
Also it's responsible for all dynamic updates, let's say a textfield
holding an algorithm parameter is changed it will notify the app
about the change and the app can recalculate the visualization.

Whitestorm is the surrounding framework for all 3D operations. It
compartmentize the complex logic into simpler modules. Also allows
several rendering tricks to make sure that our code is as crispy
as possible. It features a wide range of sublibaries like physics,
rendering, sheddering, animations, transformations.

Last but not least (more is comming), bootstrap creates the style
foundation of the web application and will help to boost development.

README features a bunch of additional Links to code that will be
needed in the near future.

The project is build with nodeJS, webpack and yarn package mgmt.
ESLint is acitvatable in the config/ directory, it's currently
deactivated, because ESLint always complains about formatting,
but not about coding errors. Which is in my opinion a distraction
from the important issues.

In the near future this code will be ported to TypeScript.