Add disable capabillity
Adds possibillity to disable parameters, instead of the form there
will be just normal labels be shown. This allows students to see
the parameters without manipulation possiblities. This should be
used for the slides in which we would like to show specific
phenomenons.

Also merged Interactive and Visualize together to have labels even
on static pictures.